fix: wrap MainContent in a suspense boundary
