+++ .gitignore
@@ -16,6 +16,9 @@ dist-ssr
 next-env.d.ts
 dist
 
+# TypeScript build cache
+.tsbuildinfo
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
