+++ src/app/manifest.json
@@ -11,7 +11,7 @@
   "lang": "en",
   "icons": [
     {
-      "src": "/favicon.svg",
+      "src": "/icon.svg",
       "sizes": "any",
       "type": "image/svg+xml",
       "purpose": "any maskable"
